move to lovely systems' docker_services
* test time reduced from 120 seconds to 14 seconds

* fixtures.py -> conftest.py for a more pytest oriented style
