+++ Pipfile
@@ -12,6 +12,7 @@ retrying = "*"
 [dev-packages]
 pytest = "*"
 pytest-docker-compose = "*"
+lovely-pytest-docker = "*"
 
 [requires]
 python_version = "3.6"
